requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langgraph>=0.0.20
langchain-mistralai>=0.0.5
streamlit>=1.29.0
pydantic>=2.5.0,<3.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
python-docx>=1.1.0
pandas>=2.1.4
numpy>=1.24.3
requests>=2.31.0
beautifulsoup4>=4.12.2
plotly>=5.17.0
streamlit-chat>=0.1.1
typing-extensions>=4.8.0
